+++ packages/ui-components/test/unit/components/UITranslationInput/utils.ts
@@ -1,31 +1,27 @@
 import { fireEvent } from '@testing-library/react';
 import type { I18nBundle } from '../../../../src/components/UITranslationInput/UITranslationButton.types';
 
-const defaultEntries = ['dummy', 'Dummy', 'dummy1', 'Dummy1', 'test', 'Test'];
+const defaultEntries = [
+    ['dummy', 'dummy text'],
+    ['Dummy'],
+    ['dummy1', 'dummy1 text'],
+    ['Dummy1', 'Dummy1 text'],
+    ['test'],
+    ['Test']
+];
 export const getBundle = (entries = defaultEntries): I18nBundle => {
     const bundle: I18nBundle = {};
     for (const entry of entries) {
-        bundle[entry] = [
+        bundle[entry[0]] = [
             {
                 key: {
-                    value: entry
+                    value: entry[0]
                 },
                 value: {
-                    value: entry
+                    value: entry[1] || entry[0]
                 }
             }
         ];
     }
     return bundle;
-};
-
-export const SELECTORS = {
-    input: '.ms-TextField',
-    button: '.ms-Button',
-    callout: '.ms-Callout'
-};
-
-export const acceptCallout = (id: string): void => {
-    const acceptBtn = document.querySelector(`#${id}-i18n-button-action-confirm`) as HTMLElement;
-    fireEvent.click(acceptBtn);
 };
